Rename to braze lambda to BrazeUsersService
